Mu4e: Improve gmail checks
Email to or from a personal Gmail address may not actually be handled by
Gmail, and regardless, email handled by Gmail must be in a Gmail
maildir, so we might as well only use that. This means we must get Gmail
maildirs from the user, because the maildir path does not have to
contain "gmail".